secure-auth/requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
flask-limiter==2.4.0
flask-session==0.4.0
redis==4.0.2
cryptography==36.0.1
flask-talisman==0.8.1
marshmallow==3.14.1
python-jose==3.3.0
flask-cors==3.0.10
flask-login==0.5.0
flask-principal==0.4.0
flask-wtf==1.0.0 